Refactor mutations and return error if login fails
